feat: 查询今日运营数据

+++ sky-pojo/src/main/java/com/sky/dto/UserReportDataDTO.java
@@ -9,5 +9,5 @@ public class UserReportDataDTO {
     // 日期
     private LocalDate date;
     // 该日期的新增用户数
-    private Long newUserCount;
+    private Integer newUserCount;
 }
